chore: 폴더 구조 정리 및 폴더명 수정
