+++ src/advanced/src/libs/findSuggestProduct.ts
