+++ src/advanced/src/libs/getStockMessage.ts
