+++ src/basic/libs/getCartQuantityDiscountRate.js
@@ -1,4 +1,4 @@
-import { DISCOUNT_RATE, DISCOUNT_THRESHOLD } from '../html/constants/constants';
+import { DISCOUNT_RATE, DISCOUNT_THRESHOLD } from '../constants/constants';
 
 export const getCartQuantityDiscountRate = (quantity) => {
   return quantity >= DISCOUNT_THRESHOLD.TOTAL ? DISCOUNT_RATE.TOTAL : 0;
