+++ src/basic/render/renderCartSummaryDetail.js
@@ -1,36 +1,37 @@
-import { findProductById } from '../utils/findProductById';
 import { isTodayTuesday } from '../utils/isTodayTuesday';
+import { findProductById } from '../libs/findProductById';
 
 // 할인 요약 내용
-export const renderCartSummaryDetail = ({ cartItems, productList, appState }) => {
+export const renderCartSummaryDetail = ({ state, appState }) => {
+  const { cartState, productState } = state;
   const { totalBeforeDiscount, totalAfterDiscount, totalProductCount, discountedProductList } = appState;
 
   const summaryDetails = document.getElementById('summary-details');
-  summaryDetails.innerHTML = '';
+  summaryDetails.innerHTML = ''; // 기존 값 초기화
+
+  let summaryText = '';
 
   // 장바구니에 상품이 존재
-  if (totalBeforeDiscount > 0) {
-    for (let i = 0; i < cartItems.length; i++) {
-      // id로 현재의 장바구니 상품 찾음 (i로 순회)
-      const curItem = findProductById(productList, cartItems[i].id);
+  if (cartState.length > 0) {
+    for (let i = 0; i < cartState.length; i++) {
+      const cartItem = cartState[i];
+      const orderCount = cartItem.count;
+      const product = findProductById(productState, cartItem.id);
 
-      // 현재 상품의 구매 수
-      const qtyElem = cartItems[i].querySelector('.quantity-number');
-      const orderCount = parseInt(qtyElem.textContent);
       // 상품 총 가격 (changedPrice - 변동된 가격, orderCount - 상품 구매 수)
-      const itemTotal = curItem.changedPrice * orderCount;
+      const itemTotal = product.changedPrice * orderCount;
 
       // 상품 이름 x 구매 수 ₩ 가격 출력
-      summaryDetails.innerHTML += /* HTML */ `
+      summaryText += /* HTML */ `
         <div class="flex justify-between text-xs tracking-wide text-gray-400">
-          <span>${curItem.name} x ${orderCount}</span>
+          <span>${product.name} x ${orderCount}</span>
           <span>₩${itemTotal.toLocaleString()}</span>
         </div>
       `;
     }
 
     // 합계 출력
-    summaryDetails.innerHTML += /* HTML */ `
+    summaryText += /* HTML */ `
       <div class="border-t border-white/10 my-3"></div>
       <div class="flex justify-between text-sm tracking-wide">
         <span>Subtotal</span>
@@ -41,7 +42,7 @@ export const renderCartSummaryDetail = ({ cartItems, productList, appState }) =>
     // 할인 정보 출력
     if (totalProductCount >= 30) {
       // 총 구매 수가 30개 이상일 때 대량 구매 할인
-      summaryDetails.innerHTML += /* HTML */ `
+      summaryText += /* HTML */ `
         <div class="flex justify-between text-sm tracking-wide text-green-400">
           <span class="text-xs">🎉 대량구매 할인 (30개 이상)</span>
           <span class="text-xs">-25%</span>
@@ -49,7 +50,7 @@ export const renderCartSummaryDetail = ({ cartItems, productList, appState }) =>
       `;
     } else if (discountedProductList.length > 0) {
       discountedProductList.forEach((item) => {
-        summaryDetails.innerHTML += /* HTML */ `
+        summaryText += /* HTML */ `
           <div class="flex justify-between text-sm tracking-wide text-green-400">
             <span class="text-xs">${item.name} (10개↑)</span>
             <span class="text-xs">-${item.discount}%</span>
@@ -61,7 +62,7 @@ export const renderCartSummaryDetail = ({ cartItems, productList, appState }) =>
     // 화요일 할인
     if (isTodayTuesday()) {
       if (totalAfterDiscount > 0) {
-        summaryDetails.innerHTML += /* HTML */ `
+        summaryText += /* HTML */ `
           <div class="flex justify-between text-sm tracking-wide text-purple-400">
             <span class="text-xs">🌟 화요일 추가 할인</span>
             <span class="text-xs">-10%</span>
@@ -70,11 +71,13 @@ export const renderCartSummaryDetail = ({ cartItems, productList, appState }) =>
       }
     }
     // 무료 배송 출력
-    summaryDetails.innerHTML += /* HTML */ `
+    summaryText += /* HTML */ `
       <div class="flex justify-between text-sm tracking-wide text-gray-400">
         <span>Shipping</span>
         <span>Free</span>
       </div>
     `;
   }
+
+  summaryDetails.innerHTML = summaryText;
 };
