+++ src/basic/render/renderCartTotalPrice.js
@@ -1,6 +1,8 @@
 // 장바구니 총 가격
-export const renderCartTotalPrice = (appState, cartTotal) => {
-  const totalDiv = cartTotal.querySelector('.text-2xl');
+export const renderCartTotalPrice = (appState) => {
+  const cartTotal = document.getElementById('cart-total');
+  const totalDiv = cartTotal?.querySelector('.text-2xl');
+
   if (totalDiv) {
     totalDiv.textContent = '₩' + Math.round(appState.totalAfterDiscount).toLocaleString();
   }
