+++ src/basic/render/renderStockMessage.js
@@ -1,7 +1,10 @@
-import { getStockMessages } from '../utils/getStockMessage';
+import { getStockMessages } from '../libs/getStockMessage';
 
 // 재고 품절 텍스트를 stockInfo에 출력
-export const renderStockMessage = (productList, stockInfo) => {
-  const stockMessage = getStockMessages(productList);
+export const renderStockMessage = (state) => {
+  const { productState } = state;
+  const stockInfo = document.getElementById('stock-status');
+  const stockMessage = getStockMessages(productState);
+
   stockInfo.textContent = stockMessage;
 };
