+++ src/basic/services/calculateBonusPoint.js
@@ -1,9 +1,9 @@
+import { BONUS_POINT, PRODUCT, MESSAGE } from '../constants/constants';
 import { isTodayTuesday } from '../utils/isTodayTuesday';
-import { findProductById } from '../utils/findProductById';
-import { PRODUCT_1, PRODUCT_2, PRODUCT_3 } from '../main.basic';
 
 // 포인트 계산
-export const calculateBonusPoint = ({ cartItems, productList, appState }) => {
+export const calculateBonusPoint = ({ state, appState }) => {
+  const { cartState } = state;
   const { totalAfterDiscount, totalProductCount } = appState;
 
   // 기본 포인트 - 총 가격 / 1000
@@ -19,55 +19,51 @@ export const calculateBonusPoint = ({ cartItems, productList, appState }) => {
   let hasMonitorArm = false;
 
   // 장바구니에 상품이 없으면 적립 포인트 요소 없앰
-  if (cartItems.length === 0) {
+  if (cartState.length === 0) {
     return { totalPoints: 0, pointsDetail: [] };
   }
 
   // 기본 포인트 출력
   if (basePoints > 0) {
     totalPoints = basePoints;
-    pointsDetail.push('기본: ' + basePoints + 'p');
+    pointsDetail.push(MESSAGE.POINT.BASE(basePoints));
   }
   // 화요일 포인트 출력
   if (isTodayTuesday()) {
     if (basePoints > 0) {
-      totalPoints = basePoints * 2;
-      pointsDetail.push('화요일 2배');
+      totalPoints = basePoints * BONUS_POINT.TUESDAY_MULTIPLIER;
+      pointsDetail.push(MESSAGE.POINT.TUESDAY);
     }
   }
 
-  for (const item of cartItems) {
-    // 아이디로 현재 상품 찾기
-    const product = findProductById(productList, item.id);
-    if (!product) continue;
-
-    // 찾은 상품으로 존재 여부 업데이트
-    if (product.id === PRODUCT_1) hasKeyboard = true;
-    if (product.id === PRODUCT_2) hasMouse = true;
-    if (product.id === PRODUCT_3) hasMonitorArm = true;
+  for (const item of cartState) {
+    // 상품 존재 여부 업데이트
+    if (item.id === PRODUCT.ID[1]) hasKeyboard = true;
+    if (item.id === PRODUCT.ID[2]) hasMouse = true;
+    if (item.id === PRODUCT.ID[3]) hasMonitorArm = true;
   }
 
   // 상품에 따른 포인트 추가
   if (hasKeyboard && hasMouse) {
-    totalPoints += 50;
-    pointsDetail.push('키보드+마우스 세트 +50p');
+    totalPoints += BONUS_POINT.KEYBOARD_MOUSE_SET;
+    pointsDetail.push(MESSAGE.POINT.KEYBOARD_MOUSE_SET);
   }
   if (hasKeyboard && hasMouse && hasMonitorArm) {
-    totalPoints += 100;
-    pointsDetail.push('풀세트 구매 +100p');
+    totalPoints += BONUS_POINT.FULL_SET;
+    pointsDetail.push(MESSAGE.POINT.FULL_SET);
   }
 
   if (totalProductCount >= 30) {
-    totalPoints += 100;
-    pointsDetail.push('대량구매(30개+) +100p');
+    totalPoints += BONUS_POINT.BULK[30];
+    pointsDetail.push(MESSAGE.POINT.BULK[30]);
   } else {
     if (totalProductCount >= 20) {
-      totalPoints += 50;
-      pointsDetail.push('대량구매(20개+) +50p');
+      totalPoints += BONUS_POINT.BULK[20];
+      pointsDetail.push(MESSAGE.POINT.BULK[20]);
     } else {
       if (totalProductCount >= 10) {
-        totalPoints += 20;
-        pointsDetail.push('대량구매(10개+) +20p');
+        totalPoints += BONUS_POINT.BULK[10];
+        pointsDetail.push(MESSAGE.POINT.BULK[10]);
       }
     }
   }
