+++ src/basic/services/calculateCartSummary.js
@@ -1,21 +1,23 @@
 import { isTodayTuesday } from '../utils/isTodayTuesday';
-import { findProductById } from '../utils/findProductById';
-import { getQuantityDiscountRate } from '../utils/getQuantityDiscountRate';
-import { getCartQuantityDiscountRate } from '../utils/getCartQuantityDiscountRate';
+import { findProductById } from '../libs/findProductById';
+import { getQuantityDiscount } from '../libs/getQuantityDiscount';
+import { getCartQuantityDiscountRate } from '../libs/getCartQuantityDiscountRate';
+import { DISCOUNT_RATE } from '../constants/constants';
 
 // 가격 및 할인율 계산
-export const calculateCartSummary = (cartItems, productList) => {
+export const calculateCartSummary = (state) => {
+  const { cartState, productState } = state;
+
   let totalBeforeDiscount = 0;
   let totalAfterDiscount = 0;
   let totalProductCount = 0;
   const discountedProductList = [];
 
-  for (let i = 0; i < cartItems.length; i++) {
-    const cartItem = cartItems[i];
-    const product = findProductById(productList, cartItem.id);
+  for (let i = 0; i < cartState.length; i++) {
+    const cartItem = cartState[i];
+    const product = findProductById(productState, cartItem.id);
 
-    const qtyElem = cartItems[i].querySelector('.quantity-number');
-    const orderCount = parseInt(qtyElem.textContent);
+    const orderCount = cartItem.count;
     const itemPrice = product.changedPrice * orderCount;
 
     // 기본 총 가격 계산
@@ -23,7 +25,7 @@ export const calculateCartSummary = (cartItems, productList) => {
     totalProductCount += orderCount;
 
     // 개별 할인 - 아이템 당 10개 이상 구매
-    const itemDiscountRate = getQuantityDiscountRate(product.id, orderCount);
+    const itemDiscountRate = getQuantityDiscount(product.id, orderCount);
     if (itemDiscountRate > 0) {
       discountedProductList.push({ name: product.name, discount: itemDiscountRate * 100 });
     }
@@ -41,7 +43,7 @@ export const calculateCartSummary = (cartItems, productList) => {
 
   // 화요일 할인 - 10% 추가 할인
   if (isTodayTuesday()) {
-    totalAfterDiscount *= 0.9;
+    totalAfterDiscount *= 1 - DISCOUNT_RATE.TUESDAY;
   }
 
   // 최종 할인율 계산
