+++ src/basic/services/updateProductSeletor.js
