+++ src/basic/states/productState.js
@@ -1,5 +1,5 @@
 import { DISCOUNT_RATE } from '../constants/constants';
-import { findProductById } from '../../utils/findProductById';
+import { findProductById } from '../libs/findProductById';
 
 export const applyFlashSale = (state, productId) => {
   const { productState } = state;
